src/data_processing/requirements.txt
pandas
mysqlconnector
pika
dotenv